requirements.txt
absl-py==0.11.0
astunparse==1.6.3
atomicwrites==1.4.0
attrs==20.3.0
cached-property==1.5.2
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-plugins==1.1.1
cligj==0.7.1
colorama==0.4.4
cycler==0.10.0
decorator==4.4.2
descartes==1.1.0
Fiona==1.8.18
flatbuffers==1.12
gast==0.3.3
GDAL==3.2.1
geopandas==0.8.2
google-auth==1.26.1
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
importlib-metadata==3.4.0
iniconfig==1.1.1
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
matplotlib==3.3.4
munch==2.5.0
networkx==2.5
numpy==1.19.5
oauthlib==3.1.0
opt-einsum==3.3.0
osmnx==1.0.1
packaging==20.9
pandas==1.2.2
Pillow==8.1.0
pluggy==0.13.1
protobuf==3.14.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pyproj==2.6.1
pytest==6.2.2
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
Rtree==0.9.7
scipy==1.6.1
Shapely==1.7.1
six==1.15.0
tb-nightly==2.5.0a20210213
tensorboard==2.4.1
tensorboard-data-server==0.3.0
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
tf-estimator-nightly==2.5.0.dev2021021301
toml==0.10.2
tqdm==4.56.2
typing-extensions==3.7.4.3
urllib3==1.26.3
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
